+++ pex/pep_508.py
@@ -53,7 +53,7 @@ class MarkerEnvironment(object):
 
         Since Pex support is (currently) restricted to:
         + interpreters: CPython and PyPy
-        + os: Linux and Mac
+        + os: Linux and Mac and Windows
 
         We can fill in most of the environment markers used in these environments in practice in the
         wild. For those we can't, we leave the markers unset (`None`).
@@ -107,6 +107,14 @@ class MarkerEnvironment(object):
             platform_machine = platform.platform.split("_", 3)[-1]
             platform_system = "Darwin"
             sys_platform = "darwin"
+        elif "win" in platform.platform:
+            os_name = "nt"
+            # E.G:
+            # + win32
+            # + win_amd64
+            platform_machine = platform.platform.split("_", 2)[-1]
+            platform_system = "Windows"
+            sys_platform = "win32"
 
         platform_python_implementation = None